// NgApp/src/app/model/InProcessBoat.ts
export class InProcessBoat {

    id: number | undefined;
    boatMasterId: number | undefined;
    boatStatus: number | undefined;
    boatName: string | undefined;
    boatType: string | undefined;
    createdDate: any | undefined;
    modifyDate: any | undefined;
    processStartDate: any | undefined;
    processEndDate: any | undefined;
    startdatetime: any | undefined;
    enddatetime: any | undefined;

}